HDM-1274 Adding a reference implementation of a storage provider plugin called tapeArchive
