+++ src/main/resources/application-dev.properties
@@ -13,14 +13,14 @@ s3_secret_key = 0/0/0
 s3_bucket = bucket_name
 s3_region = east
 
-sda_queued_proc_period = 500
-sda_timer_period = 5s
-sda_processor_bean = defaultProcessor
-sda_cache_id = cacheId
-sda_hostname = localhost
-sda_port = 8101
-sda_servicename = cacheManager.cgi/cache/
-sda_fixityservice = cacheManager.cgi/checksum/
+tapeArchive_queued_proc_period = 500
+tapeArchive_timer_period = 5s
+tapeArchive_processor_bean = defaultProcessor
+tapeArchive_cache_id = cacheId
+tapeArchive_hostname = localhost
+tapeArchive_port = 8101
+tapeArchive_servicename = tapeArchiveService/cache/
+tapeArchive_fixityservice = tapeArchiveService/checksum/
 
 management.port = 10001
