--- src/main/resources/wiremock/mappings/tape_service_response.json
@@ -1,13 +1,0 @@
-{
-    "request": {
-        "method": "GET",
-        "url": "/tapeArchiveService/cache/some_media_file_01_access.mp4"
-    },
-    "response": {
-        "status": 200,
-        "body": "{\n \"url\":\"/download_cache/some_media_file_01_access.mp4\",\"error\":0,\"message\":\"file is present\"\n}",
-        "headers" : {
-          "Content-Type" : "application/json"
-        }
-    }
-}
